[bugFix] Исправлен вывод и переполнение коэффициентов (int -> long long)

+++ poly.h
@@ -11,8 +11,8 @@
 
 typedef struct polyElement {
     char variable;
-    int coeff;
-    int degree;
+    long long coeff;
+    long long degree;
     struct polyElement* next;
 } polyElement;
 
@@ -27,14 +27,14 @@ void init_structs();
 void set_polynom(char varname, polyElement* polynom);
 polyElement* search_polynom(char varname);
 
-polyElement* create(char variable, int coeff, int degree);
+polyElement* create(char variable, long long coeff, long long degree);
 polyElement* mult(polyElement* elem1, polyElement* elem2);
 polyElement* polydegree(polyElement* elem, polyElement* degree);
-polyElement* degree(polyElement* elem, int degree);
+polyElement* degree(polyElement* elem, long long degree);
 polyElement* summary(polyElement* elem1, polyElement* elem2);
 polyElement* diff(polyElement* elem1, polyElement* elem2);
-void mult_coeff(polyElement* elem, int coeff);
-int calc_degree(int value, int deg);
+void mult_coeff(polyElement* elem, long long coeff);
+int calc_degree(long long value, long long deg);
 polyElement* unary(polyElement* elem);
 polyElement* copy_polynoms(polyElement* dest, polyElement* src);
